project

+++ src/by/it/obmetko/project/java/DAO/beens/Book.java
@@ -6,11 +6,11 @@ public class Book {
     private int id;
     private String name;
     private String author;
-    private double price;
+    private int price;
     private int Category_id;
 
 
-    public Book(int id, String name, String author, double price, int category_id) {
+    public Book(int id, String name, String author, int price, int category_id) {
         this.id = id;
         this.name = name;
         this.author = author;
@@ -42,11 +42,11 @@ public class Book {
         this.author = author;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
